Fixes issue with properly applying convolution kernel

+++ galaxy2galaxy/data_generators/galsim_utils.py
@@ -154,7 +154,7 @@ def draw_and_encode_stamp(gal, psf, stamp_size, pixel_scale, attributes=None):
 
     # We draw the pixel image of the convolved image
     im = gal.drawImage(nx=stamp_size, ny=stamp_size, scale=pixel_scale,
-                       method='no_pixel', use_true_center=False).array
+                       method='no_pixel', use_true_center=False).array.astype('float32')
 
     # Draw the Fourier domain image of the galaxy, using x2 zero padding,
     # and x2 subsampling
